fgeo: Easily Install and Load Packages of ForestGEO

'fgeo' installs and load multiple R packages of ForestGEO in a single step. Each individual package is a self contained module. Together, all the packages in 'fgeo' provide a wide renge of tools to analyze forest diversity and dynamics. The modular structure of 'fgeo' makes it easy to add and maintain more tools as they become available. Learn more about the 'fgeo' package at <http://forestgeo.github.io/fgeo>; and learn more about ForestGEO at <http://www.forestgeo.si.edu/>.

See also

Useful links:

http://forestgeo.github.io/fgeo

https://github.com/forestgeo/fgeo

Report bugs at https://github.com/forestgeo/fgeo/issues