Index fgeo's elements.
Source: R/fgeo_index.R

fgeo_index() created a table of some component of fgeo for each core package of fgeo. The other functions are convenient wrappers.

fgeo_index(.f, nm, ...)

fgeo_index_functions(keep_reexported = FALSE)

fgeo_index_packages()

fgeo_index_datasets()

Arguments

.f
A function to map to each core package of fgeo.

nm
A name for the column storing the results of .f.

...
Additional arguments passed to .f.

keep_reexported
Logical; TRUE includes reexported functions.

Value

A dataframe.

Acknowledgements

Jim Hester adviced on implementation details of fgeo_index_functions(); and Nathan Werth adviced on implementation details of fgeo_index_datasets(). detail.

Examples

head(fgeo_index_functions())

#>   package                       fun
#> 3    fgeo check_if_pkg_can_be_found
#> 5    fgeo          confirm_conflict
#> 6    fgeo               core_loaded
#> 7    fgeo             core_unloaded
#> 8    fgeo             external_funs
#> 9    fgeo               fgeo_attach

head(fgeo_index_packages())

#>           package                                           Title
#> 1           bciex Forest Dynamics Data from Barro Colorado Island
#> 2            fgeo   Easily Install and Load Packages of ForestGEO
#> 3  fgeo.abundance  R packages for the Analysis of Forest Dynamics
#> 4 fgeo.demography                                Demography Tools
#> 5    fgeo.habitat                                Habitat Analyses
#> 6        fgeo.map                            Map ForestGEO's Data