Explore the documentation of fgeo.

Explore the documentation of fgeo.

fgeo_docs(pattern = NULL, ..., package = fgeo::fgeo_core())

Arguments

pattern
Character string of length one to match an entry in packages documentation.

...
Bare names of the columns to select. Valid names are package, name, title, topic, type, alias, keyword, concept. If no name is given, then all names are returned.

package
A character vector with the names of packages to search through, or NULL in which case all available packages in the library trees specified by lib.loc are searched. Defaults to search only the core packages of fgeo.

Value

A dataframe.

See also

Examples

# Filter rows with a matching pattern.
fgeo_docs("predicate")

#> # A tibble: 8 x 7
#>   package  title                     topic   type  alias   keyword concept     
#>   <chr>    <chr>                     <chr>   <chr> <chr>   <chr>   <chr>       
#> 1 fgeo.to~ Detect and flag based on~ flag_i~ help  flag_i~ intern~ functions f~
#> 2 fgeo.to~ Detect and flag based on~ flag_i~ help  flag_i~ intern~ functions t~
#> 3 fgeo.to~ Detect and flag based on~ flag_i~ help  flag_i~ intern~ predicates. 
#> 4 fgeo.to~ Detect and flag based on~ flag_i~ help  detect~ intern~ functions f~
#> 5 fgeo.to~ Detect and flag based on~ flag_i~ help  detect~ intern~ functions t~
#> 6 fgeo.to~ Detect and flag based on~ flag_i~ help  detect~ intern~ predicates. 
#> 7 fgeo.to~ Predicates to detect and~ is_mul~ help  is_mul~ intern~ general pre~
#> 8 fgeo.to~ Predicates to detect and~ is_mul~ help  is_dup~ intern~ general pre~

# Select specific columns
fgeo_docs("predicate", concept, topic, title)

#> # A tibble: 4 x 3
#>   concept             topic      title                                         
#>   <chr>               <chr>      <chr>                                         
#> 1 functions for deve~ flag_if_g~ Detect and flag based on a predicate applied ~
#> 2 functions to check~ flag_if_g~ Detect and flag based on a predicate applied ~
#> 3 predicates.         flag_if_g~ Detect and flag based on a predicate applied ~
#> 4 general predicates  is_multip~ Predicates to detect and flag duplicated and ~

# Exclude specific columns
fgeo_docs("abundance", -package)

#> # A tibble: 7 x 6
#>   title                  topic     type  alias     keyword concept             
#>   <chr>                  <chr>     <chr> <chr>     <chr>   <chr>               
#> 1 Abundance and basal a~ abundance help  abundance NA      NA                  
#> 2 Abundance and basal a~ abundance help  basal_ar~ NA      NA                  
#> 3 Create tables of abun~ abundanc~ help  abundanc~ NA      NA                  
#> 4 Create tables of abun~ abundanc~ help  basal_ar~ NA      NA                  
#> 5 fgeo.abundance: Calcu~ fgeo.abu~ help  fgeo.abu~ intern~ NA                  
#> 6 fgeo.abundance: Calcu~ fgeo.abu~ help  fgeo.abu~ intern~ NA                  
#> 7 A version of ctfs::ab~ abundanc~ help  abundanc~ intern~ functions from http~