Get help with fgeo.

fgeo_help() finds all fgeo help files. You can refine your search directly on the viewer panel of RStudio or by passing a string to fgeo_help() (e.g. fgeo("datasets")).

fgeo_help_adin() runs fgeo_help() from the RStudio Addins menu or by associating a keyboard shortcut to it (see Tools > Modify Keyboard Shortcuts...), for example, Ctr/Cmd + Shift + Alt + F1.

fgeo_help(pattern = "", package = fgeo::fgeo_core())

fgeo_help_addin()

Arguments

pattern
A character string to be matched in the name, alias or title of a topic's documentation. Defaults to matching anything, which retrieves all the documentation of fgeo packages.

package
A character string giving the name or one or more core-packages of fgeo.

Value

The results are returned in a list object of class "hsearch", which has a print method for nicely formatting the results of the query.

See also

Examples

if (interactive()) fgeo_help()
if (interactive()) fgeo_help("fgeo", package = "fgeo")

results <- fgeo_help("fgeo", package = "fgeo.tool")
if (interactive()) results
str(results)

#> List of 9
#>  $ pattern    : chr "fgeo"
#>  $ fields     : chr [1:3] "alias" "concept" "title"
#>  $ type       : chr "regexp"
#>  $ agrep      : NULL
#>  $ ignore.case: logi TRUE
#>  $ types      : chr [1:3] "vignette" "demo" "help"
#>  $ package    : chr "fgeo.tool"
#>  $ lib.loc    : chr "C:/Users/LeporeM/Documents/R/R-3.5.1/library"
#>  $ matches    :'data.frame':	24 obs. of  9 variables:
#>   ..$ Topic  : chr [1:24] "add_status_tree" "add_status_tree" "add_subquad" "add_var" ...
#>   ..$ Title  : chr [1:24] "Add column 'status_tree' based on the status of all stems of each tree." "Add column 'status_tree' based on the status of all stems of each tree." "Add column 'subquadrat' based on 'qx' and 'qy' coordinates." "Add columns 'lx/ly'; 'QX/QY'; 'index'; 'col/row'; and 'hectindex', 'quad'." ...
#>   ..$ Name   : chr [1:24] "add_status_tree" "add_status_tree" "add_subquad" "add_var" ...
#>   ..$ ID     : chr [1:24] "4/2" "4/2" "4/3" "4/4" ...
#>   ..$ Package: chr [1:24] "fgeo.tool" "fgeo.tool" "fgeo.tool" "fgeo.tool" ...
#>   ..$ LibPath: chr [1:24] "C:/Users/LeporeM/Documents/R/R-3.5.1/library/fgeo.tool" "C:/Users/LeporeM/Documents/R/R-3.5.1/library/fgeo.tool" "C:/Users/LeporeM/Documents/R/R-3.5.1/library/fgeo.tool" "C:/Users/LeporeM/Documents/R/R-3.5.1/library/fgeo.tool" ...
#>   ..$ Type   : chr [1:24] "help" "help" "help" "help" ...
#>   ..$ Field  : chr [1:24] "concept" "concept" "concept" "concept" ...
#>   ..$ Entry  : chr [1:24] "functions for fgeo census." "functions for fgeo vft." "c(\"functions for fgeo vft.\\n\", \"\\n\", \"  ```R\\n\", \"  First column is 0    First column is 1\\n\", \"  "| __truncated__ "functions for fgeo census." ...
#>  - attr(*, "class")= chr "hsearch"