fgeo: Install and Load Multiple Packages in One Step

'fgeo' installs and loads multiple R packages of ForestGEO in a single step. Each individual package is a self-contained module. Together, all the packages in 'fgeo' provide a wide range of tools to analyze forest diversity and dynamics. The modular structure of 'fgeo' makes it easy to add and maintain more tools as they become available. Learn more about the 'fgeo' package at <http://forestgeo.github.io/fgeo>; and learn more about ForestGEO at <http://www.forestgeo.si.edu/>.

See also