Explore the documentation of fgeo.

Explore the documentation of fgeo.

fgeo_docs(pattern = NULL, ..., package = fgeo::fgeo_core())

Arguments

pattern
Character string of length one to match an entry in packages documentation.

...
Bare names of the columns to select. Valid names are package, name, title, topic, type, alias, keyword, concept. If no name is given, then all names are returned.

package
A character vector with the names of packages to search through, or NULL in which case all available packages in the library trees specified by lib.loc are searched. Defaults to search only the core packages of fgeo.

Value

A dataframe.

Examples

# Filter rows with a matching pattern.
fgeo_docs("predicate")

#> # A tibble: 11 x 7
#>    package  title                   topic    type  alias    keyword concept    
#>    <chr>    <chr>                   <chr>    <chr> <chr>    <chr>   <chr>      
#>  1 fgeo.ba~ Apply a predicate func~ detect_~ help  detect_~ <NA>    <NA>       
#>  2 fgeo.ba~ Detect and extract mat~ extract~ help  extract~ <NA>    predicates.
#>  3 fgeo.ba~ Detect and extract mat~ extract~ help  detect_~ <NA>    predicates.
#>  4 fgeo.ba~ Predicates to detect a~ is_mult~ help  is_mult~ <NA>    predicates.
#>  5 fgeo.ba~ Predicates to detect a~ is_mult~ help  is_dupl~ <NA>    predicates.
#>  6 fgeo.to~ Detect and flag based ~ flag_if~ help  flag_if~ <NA>    functions ~
#>  7 fgeo.to~ Detect and flag based ~ flag_if~ help  flag_if~ <NA>    functions ~
#>  8 fgeo.to~ Detect and flag based ~ flag_if~ help  flag_if~ <NA>    predicates.
#>  9 fgeo.to~ Detect and flag based ~ flag_if~ help  detect_~ <NA>    functions ~
#> 10 fgeo.to~ Detect and flag based ~ flag_if~ help  detect_~ <NA>    functions ~
#> 11 fgeo.to~ Detect and flag based ~ flag_if~ help  detect_~ <NA>    predicates.

# Select specific columns
fgeo_docs("predicate", concept, topic, title)

#> # A tibble: 6 x 3
#>   concept            topic         title                                       
#>   <chr>              <chr>         <chr>                                       
#> 1 <NA>               detect_if     Apply a predicate function to a column of a~
#> 2 predicates.        extract_inse~ Detect and extract matching strings - ignor~
#> 3 predicates.        is_multiple   Predicates to detect and flag duplicated an~
#> 4 functions for dev~ flag_if_group Detect and flag based on a predicate applie~
#> 5 functions to chec~ flag_if_group Detect and flag based on a predicate applie~
#> 6 predicates.        flag_if_group Detect and flag based on a predicate applie~

# Exclude specific columns
fgeo_docs("abundance", -package)

#> # A tibble: 7 x 6
#>   title                           topic      type  alias        keyword concept
#>   <chr>                           <chr>      <chr> <chr>        <chr>   <chr>  
#> 1 Abundance and basal area, opti~ abundance  help  abundance    <NA>    <NA>   
#> 2 Abundance and basal area, opti~ abundance  help  basal_area   <NA>    <NA>   
#> 3 Create tables of abundance and~ abundance~ help  abundance_b~ <NA>    <NA>   
#> 4 Create tables of abundance and~ abundance~ help  basal_area_~ <NA>    <NA>   
#> 5 fgeo.abundance: Calculate Abun~ fgeo.abun~ help  fgeo.abunda~ intern~ <NA>   
#> 6 fgeo.abundance: Calculate Abun~ fgeo.abun~ help  fgeo.abunda~ intern~ <NA>   
#> 7 Count and basal area of neighb~ neighbor   help  abundance_n~ <NA>    <NA>