Search fgeo for documentation matching a given character string.

Use this function with a string to match the name of a topic (generally the name of a function or dataset), or text from the title or alias sections of the topic's documentation.

fgeo_help(pattern = "", package = fgeo::fgeo_core())

Arguments

pattern
A character string to be matched in the name, alias or title of a topic's documentation. Defaults to matching anything, which retrieves all the documentation of fgeo packages.

package
A character string giving the name or one or more core-packages of fgeo.

Value

The results are returned in a list object of class "hsearch", which has a print method for nicely formatting the results of the query.

See also

Examples

if (interactive()) fgeo_help()
if (interactive()) fgeo_help("fgeo", package = "fgeo")

results <- fgeo_help("fgeo", package = "fgeo.tool")
if (interactive()) results
str(results)

#> List of 9
#>  $ pattern    : chr "fgeo"
#>  $ fields     : chr [1:3] "alias" "concept" "title"
#>  $ type       : chr "regexp"
#>  $ agrep      : NULL
#>  $ ignore.case: logi TRUE
#>  $ types      : chr [1:3] "vignette" "demo" "help"
#>  $ package    : chr "fgeo.tool"
#>  $ lib.loc    : chr "C:/Users/LeporeM/Documents/R/R-3.5.1/library"
#>  $ matches    :'data.frame':	15 obs. of  9 variables:
#>   ..$ Topic  : chr [1:15] "add_status_tree" "add_status_tree" "add_subquad" "add_var" ...
#>   ..$ Title  : chr [1:15] "Define the status of a tree based on the status of its stems." "Define the status of a tree based on the status of its stems." "Add a quadrat variable to a dataframe based based on qx and qy coordinates." "Add columns to a ForestGEO dataframe to position stems in a forest plot." ...
#>   ..$ Name   : chr [1:15] "add_status_tree" "add_status_tree" "add_subquad" "add_var" ...
#>   ..$ ID     : chr [1:15] "1/1" "1/1" "1/2" "1/3" ...
#>   ..$ Package: chr [1:15] "fgeo.tool" "fgeo.tool" "fgeo.tool" "fgeo.tool" ...
#>   ..$ LibPath: chr [1:15] "C:/Users/LeporeM/Documents/R/R-3.5.1/library/fgeo.tool" "C:/Users/LeporeM/Documents/R/R-3.5.1/library/fgeo.tool" "C:/Users/LeporeM/Documents/R/R-3.5.1/library/fgeo.tool" "C:/Users/LeporeM/Documents/R/R-3.5.1/library/fgeo.tool" ...
#>   ..$ Type   : chr [1:15] "help" "help" "help" "help" ...
#>   ..$ Field  : chr [1:15] "concept" "concept" "concept" "concept" ...
#>   ..$ Entry  : chr [1:15] "functions for fgeo census." "functions for fgeo vft." "c(\"functions for fgeo vft.\\n\", \"\\n\", \"  ```R\\n\", \"  First column is 0    First column is 1\\n\", \"  "| __truncated__ "functions for fgeo census." ...
#>  - attr(*, "class")= chr "hsearch"