Get help with fgeo.

fgeo_help() finds all fgeo help files. You can refine your search directly on the viewer panel of RStudio or by passing a string to fgeo_help() (e.g. fgeo_help("datasets")).

fgeo_help_adin() runs fgeo_help() from the RStudio Addins menu or by associating a keyboard shortcut to it (see Tools > Modify Keyboard Shortcuts...), for example, Ctr/Cmd + Shift + Alt + F1.

fgeo_help(pattern = "", package = NULL, rebuild = TRUE, ...)

fgeo_help_addin()

Arguments

pattern
A character string to be matched in the name, alias or title of a topic's documentation. Defaults to matching anything, which retrieves all the documentation of fgeo packages.

package
A character string giving the name or one or more core-packages of fgeo.

rebuild
a logical indicating whether the help database should be rebuilt. This will be done automatically if lib.loc or the search path is changed, or if package is used and a value is not found.

...
Other arguments passed to utils::help.search().

Value

The results are returned in a list object of class "hsearch", which has a print method for nicely formatting the results of the query.

See also

Examples

if (interactive()) fgeo_help()
as_tibble(fgeo_help()$matches)

#> # A tibble: 426 x 9
#>    Topic   Title       Name    ID    Package LibPath      Type  Field Entry     
#>    <chr>   <chr>       <chr>   <chr> <chr>   <chr>        <chr> <chr> <chr>     
#>  1 abunda~ Abundance ~ abunda~ 1/1   fgeo.a~ C:/Users/Le~ help  alias abundance 
#>  2 abunda~ Abundance ~ abunda~ 1/1   fgeo.a~ C:/Users/Le~ help  alias basal_area
#>  3 abunda~ Abundance ~ abunda~ 1/1   fgeo.a~ C:/Users/Le~ help  conc~ functions~
#>  4 abunda~ Abundance ~ abunda~ 1/1   fgeo.a~ C:/Users/Le~ help  Title Abundance~
#>  5 abunda~ Create tab~ abunda~ 1/2   fgeo.a~ C:/Users/Le~ help  alias abundance~
#>  6 abunda~ Create tab~ abunda~ 1/2   fgeo.a~ C:/Users/Le~ help  alias basal_are~
#>  7 abunda~ Create tab~ abunda~ 1/2   fgeo.a~ C:/Users/Le~ help  conc~ functions~
#>  8 abunda~ Create tab~ abunda~ 1/2   fgeo.a~ C:/Users/Le~ help  Title Create ta~
#>  9 demogr~ Recruitmen~ demogr~ 1/3   fgeo.a~ C:/Users/Le~ help  alias demograph~
#> 10 demogr~ Recruitmen~ demogr~ 1/3   fgeo.a~ C:/Users/Le~ help  alias recruitme~
#> # ... with 416 more rows

if (interactive()) fgeo_help("stem", package = "fgeo.x")