Get help with fgeo.

fgeo_help() finds all fgeo help files. You can refine your search directly on the viewer panel of RStudio or via a string passed as the first argument to fgeo_help().

fgeo_help(pattern = "", package = NULL, rebuild = TRUE, ...)

Arguments

pattern
A character string to be matched in the name, alias or title of a topic's documentation. Defaults to matching everything, which retrieves all the documentation of fgeo packages.

package
A character string giving the name of one or more packages to limit the search, or NULL to search all fgeo packages.

rebuild
a logical indicating whether the help database should be rebuilt. This will be done automatically if lib.loc or the search path is changed, or if package is used and a value is not found.

...
Other arguments passed to utils::help.search().

Value

The results are returned in a list object of class "hsearch", which has a print method for nicely formatting the results of the query.

See also

Examples

if (interactive()) {
  fgeo_help()
}

dplyr::as_tibble(fgeo_help()$matches)

#> # A tibble: 327 x 9
#>    Topic  Title           Name   ID    Package LibPath        Type  Field Entry 
#>    <chr>  <chr>           <chr>  <chr> <chr>   <chr>          <chr> <chr> <chr> 
#>  1 datas~ Access data vi~ datas~ 1/1   fgeo.x  C:/Users/Lepo~ help  alias datas~
#>  2 datas~ Access data vi~ datas~ 1/1   fgeo.x  C:/Users/Lepo~ help  alias eleva~
#>  3 datas~ Access data vi~ datas~ 1/1   fgeo.x  C:/Users/Lepo~ help  alias habit~
#>  4 datas~ Access data vi~ datas~ 1/1   fgeo.x  C:/Users/Lepo~ help  alias stem5 
#>  5 datas~ Access data vi~ datas~ 1/1   fgeo.x  C:/Users/Lepo~ help  alias stem6 
#>  6 datas~ Access data vi~ datas~ 1/1   fgeo.x  C:/Users/Lepo~ help  alias tree5 
#>  7 datas~ Access data vi~ datas~ 1/1   fgeo.x  C:/Users/Lepo~ help  alias tree6 
#>  8 datas~ Access data vi~ datas~ 1/1   fgeo.x  C:/Users/Lepo~ help  alias vft_4~
#>  9 datas~ Access data vi~ datas~ 1/1   fgeo.x  C:/Users/Lepo~ help  alias tree6~
#> 10 datas~ Access data vi~ datas~ 1/1   fgeo.x  C:/Users/Lepo~ help  conc~ datas~
#> # ... with 317 more rows

if (interactive()) {
  fgeo_help("stem", package = "fgeo.x")
}